+++ attributes/default.rb
@@ -4,4 +4,5 @@ include_attribute 'all-datacenter-attributes'
 cookbook_name = 'winbind'
 recipe_name = 'default'
 
-# placeholder
+include_attribute 'all-datacenter-attributes::realm'	# AD centralized settings
+